Split values.rs into separate pieces
values.rs was confusing as to why Visit was implemented multiple times for different structs. Split the two implementations into separate files with better names.

Also removed an unnecessary generic.

Also add some basic builder tests.
