+++ src/native/common_schema/user_events_cs.rs
@@ -1,3 +1,4 @@
+use crate::values::{*, event_values::*};
 use crate::error::EtwError;
 use crate::values::*;
 use eventheader::*;
@@ -20,12 +21,12 @@ pub(crate) struct CommonSchemaPartCBuilder<'a> {
 }
 
 impl<'a> CommonSchemaPartCBuilder<'a> {
-    fn make_visitor(eb: &'a mut EventBuilder) -> VisitorWrapper<CommonSchemaPartCBuilder<'a>> {
-        VisitorWrapper::from(CommonSchemaPartCBuilder { eb })
+    fn make_visitor(eb: &'a mut EventBuilder) -> EventBuilderVisitorWrapper<CommonSchemaPartCBuilder<'a>> {
+        EventBuilderVisitorWrapper::from(CommonSchemaPartCBuilder { eb })
     }
 }
 
-impl<T> AddFieldAndValue<T> for CommonSchemaPartCBuilder<'_> {
+impl AddFieldAndValue for CommonSchemaPartCBuilder<'_> {
     fn add_field_value(&mut self, fv: &FieldAndValue) {
         let mut field_name: &'static str = fv.field_name;
 
@@ -34,7 +35,7 @@ impl<T> AddFieldAndValue<T> for CommonSchemaPartCBuilder<'_> {
             assert!(matches!(fv.value, ValueTypes::v_str(_)));
         }
 
-        <&mut EventBuilder as AddFieldAndValue<EventBuilder>>::add_field_value(
+        <&mut EventBuilder as AddFieldAndValue>::add_field_value(
             &mut self.eb,
             &FieldAndValue {
                 field_name,
@@ -167,7 +168,7 @@ impl crate::native::EventWriter<CommonSchemaProvider> for CommonSchemaProvider {
         _timestamp: SystemTime,
         _activity_id: &[u8; 16],
         _related_activity_id: &[u8; 16],
-        _fields: &'b [crate::values::FieldValueIndex],
+        _fields: &'b [crate::values::span_values::FieldValueIndex],
         _level: &tracing_core::Level,
         _keyword: u64,
         _event_tag: u32,
@@ -182,7 +183,7 @@ impl crate::native::EventWriter<CommonSchemaProvider> for CommonSchemaProvider {
         start_stop_times: (std::time::SystemTime, std::time::SystemTime),
         _activity_id: &[u8; 16],
         _related_activity_id: &[u8; 16],
-        fields: &'b [crate::values::FieldValueIndex],
+        fields: &'b [crate::values::span_values::FieldValueIndex],
         level: &tracing_core::Level,
         keyword: u64,
         event_tag: u32,
@@ -277,9 +278,7 @@ impl crate::native::EventWriter<CommonSchemaProvider> for CommonSchemaProvider {
                 let mut pfv = CommonSchemaPartCBuilder { eb: eb.deref_mut() };
 
                 for f in fields {
-                    <CommonSchemaPartCBuilder<'_> as AddFieldAndValue<
-                        CommonSchemaPartCBuilder<'_>,
-                    >>::add_field_value(
+                    <CommonSchemaPartCBuilder<'_> as AddFieldAndValue>::add_field_value(
                         &mut pfv,
                         &FieldAndValue {
                             field_name: f.field,
