+++ src/native/noop.rs
@@ -45,7 +45,7 @@ impl crate::native::EventWriter<Provider> for Provider {
         _timestamp: SystemTime,
         _activity_id: &[u8; 16],
         _related_activity_id: &[u8; 16],
-        _fields: &'b [crate::values::FieldValueIndex],
+        _fields: &'b [crate::values::span_values::FieldValueIndex],
         _level: &tracing_core::Level,
         _keyword: u64,
         _event_tag: u32,
@@ -60,7 +60,7 @@ impl crate::native::EventWriter<Provider> for Provider {
         _start_stop_times: (std::time::SystemTime, std::time::SystemTime),
         _activity_id: &[u8; 16],
         _related_activity_id: &[u8; 16],
-        _fields: &'b [crate::values::FieldValueIndex],
+        _fields: &'b [crate::values::span_values::FieldValueIndex],
         _level: &tracing_core::Level,
         _keyword: u64,
         _event_tag: u32,
