+++ src/native/user_events.rs
@@ -1,3 +1,5 @@
+use crate::values::{*, event_values::*};
+use crate::statics::GLOBAL_ACTIVITY_SEED;
 use crate::error::EtwError;
 use crate::statics::GLOBAL_ACTIVITY_SEED;
 use crate::values::*;
@@ -20,7 +22,7 @@ static mut ETW_META_PTR: *const crate::_details::EventMetadata = core::ptr::null
 
 thread_local! {static EBW: std::cell::RefCell<EventBuilder>  = RefCell::new(EventBuilder::new());}
 
-impl<T> AddFieldAndValue<T> for &'_ mut eventheader_dynamic::EventBuilder {
+impl AddFieldAndValue for &'_ mut eventheader_dynamic::EventBuilder {
     fn add_field_value(&mut self, fv: &FieldAndValue) {
         match fv.value {
             ValueTypes::None => (),
@@ -175,7 +177,7 @@ impl crate::native::EventWriter<Provider> for Provider {
         timestamp: SystemTime,
         activity_id: &[u8; 16],
         related_activity_id: &[u8; 16],
-        fields: &'b [crate::values::FieldValueIndex],
+        fields: &'b [crate::values::span_values::FieldValueIndex],
         level: &tracing_core::Level,
         keyword: u64,
         event_tag: u32,
@@ -207,7 +209,7 @@ impl crate::native::EventWriter<Provider> for Provider {
             );
 
             for f in fields {
-                <&mut EventBuilder as AddFieldAndValue<EventBuilder>>::add_field_value(
+                <&mut EventBuilder as AddFieldAndValue>::add_field_value(
                     &mut eb.deref_mut(),
                     &FieldAndValue {
                         field_name: f.field,
@@ -238,7 +240,7 @@ impl crate::native::EventWriter<Provider> for Provider {
         start_stop_times: (std::time::SystemTime, std::time::SystemTime),
         activity_id: &[u8; 16],
         related_activity_id: &[u8; 16],
-        fields: &'b [crate::values::FieldValueIndex],
+        fields: &'b [crate::values::span_values::FieldValueIndex],
         level: &tracing_core::Level,
         keyword: u64,
         event_tag: u32,
@@ -271,7 +273,7 @@ impl crate::native::EventWriter<Provider> for Provider {
             );
 
             for f in fields {
-                <&mut EventBuilder as AddFieldAndValue<EventBuilder>>::add_field_value(
+                <&mut EventBuilder as AddFieldAndValue>::add_field_value(
                     &mut eb.deref_mut(),
                     &FieldAndValue {
                         field_name: f.field,
@@ -347,7 +349,7 @@ impl crate::native::EventWriter<Provider> for Provider {
                 0,
             );
 
-            let mut visitor = VisitorWrapper::from(eb.deref_mut());
+            let mut visitor = EventBuilderVisitorWrapper::from(eb.deref_mut());
             event.record(&mut visitor);
 
             let _ = eb.write(
